requirements.txt
numpy==1.21.4
onnxruntime==1.9.0
Pillow==9.0.1
requests==2.27.1
torch==1.10.1
torchvision==0.11.2
